removed duplicates
